fixing test file

+++ tests/test_rmq.sh
@@ -1,16 +1,13 @@
 #!/bin/bash
 
 # Start backend
-python3 backend/backend.py &
-BACKEND_PID=$!
+# python3 ../backend/backend.py & BACKEND_PID=$!
 
 # Start dbworker
-python3 database/dbworker.py &
-DBWORKER_PID=$!
+# python3 ../database/dbworker.py & DBWORKER_PID=$!
 
 # Start middleware
-node frontend/middleware.js &
-MIDDLEWARE_PID=$!
+node ../frontend/middleware.js & MIDDLEWARE_PID=$!
 
 # Wait for services to start
 sleep 5
@@ -22,6 +19,12 @@ RESPONSE=$(curl -s -X POST http://localhost:3000/api/form-submit -H "Content-Typ
 echo "Response: $RESPONSE"
 
 # Kill all processes
-kill $BACKEND_PID
-kill $DBWORKER_PID
-kill $MIDDLEWARE_PID
+if [ -z $BACKEND_BID ]; then
+	kill $BACKEND_PID
+fi
+if [ -z $DBWORKER_PID ]; then
+	kill $DBWORKER_PID
+fi
+if [ -z $MIDDLEWARE_PID ]; then
+	kill $MIDDLEWARE_PID
+fi
